+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 - In default scopes, replace offline per offline_access (The standard value)
 - Add `--ttl` and `--renewAt` parameters on `client` and `client-nui` subcommand to exercise token renewal.
+- Introspection endpoint is now fetched from discovery mechanism.